CI: Temporary fix for picotool fetching develop branch.
Pico SDK 2.1.1 was released referencing the develop branch of picotool,
which was later fixed in a commit post 2.1.1 and only available on the
master branch.

This change vendors the upstream patch against master and backports it
to the 2.1.1 tree in MicroPython. This is a fragile, temporary fix
to get this build out of the door- it *will* break when MicroPython
bump Pico SDK.

TODO: Investigate fetching and installing picotool standalone.

See:

https://github.com/raspberrypi/picotool/issues/256